requirements.txt
streamlit
flask
flask-cors
easyocr
pillow
